feat: add Codestral on Vertex for Code Completions
We are introducing Codestral as the new model for Code Completions.
We decided to use Vertex as the provider for Codestral.
See: https://gitlab.com/gitlab-org/gitlab/-/issues/470171

LiteLLM SDK is used to implement this since it supports Codestral-on-Vertex.
This is also implemented with the Agent registry so that the prompt
is set on the AI Gateway side.

Address review feedback

Resolve lint errors, refactor for code quality

Add tests, refactor for code quality

Update vertex-on-codestral temp

Set the vertex location for codestral prompt