feat: add Codestral on Vertex for Code Completions
We are introducing Codestral as the new model for Code Completions.
We decided to use Vertex as the provider for Codestral.
See: https://gitlab.com/gitlab-org/gitlab/-/issues/470171

LiteLLM SDK is used to implement this since it supports Codestral-on-Vertex.
This is also implemented with the Agent registry so that the prompt
is set on the AI Gateway side.