fix(deps): force mise to compile Python
Currently poetry.toml configures virtual environments with
`always-copy = true`, but this doesn't work with the precompiled
Python packages installed by mise because the `python3` binary is
symlinked to a shared library with a relative path.

To fix this just compile Python from scratch to ensure the binary is
symlinked to an absolute path.

Alternatively we could:

1. Drop the `always-copy` setting.
2. Fix Poetry so that it also copies over the shared library.

Relates to https://github.com/mise-plugins/mise-poetry/issues/16