feat: blocks logging for requests unless explicit
expanded_ai_logging feature flag support was added to AIGW so we can
control when detailed information is logged, but only a few selected
endpoints where actually using it. This creates a logger wrapper and
replaces all endpoints to not log debug info when the feature flag is
not passed or when the env var is not set.

Changelog: added