refactor: separate initialization from input handling in ReAct agent
Currently, the `ReActAgent` class is not really working as a `Runnable`,
in that the input is not provided at the time of execution (when calling
`ainvoke` or `astream`) but when initiializing the agent object. Aside
from breaking conventions, this flow makes it harder to handle auth,
because everywhere else in the code we delegate checking Unit Primitives
to  `BasePromptRegistry.get_on_behalf`, so we're certain that the prompt is
properly authenticated by the time we run it; whereas with the ReAct agent,
we have to call the barebones `get` method of the registry, create a prompt
object that the user may or may not have access to, manually perform
authorization, and then proceed to execution. This change removes the input
needed to assemble the agent, the `messages`, from `ReActInputs`, so we can
fully separate building and authorizing the agent, and thus we can properly
send the inputs to the execution methods