fix: treat `ModelMetadata` as a model instead of a dict
In
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/1744
we used `model_dump` to create a kwargs dict to pass to `get_on_behalf`,
in order to exclude `None` values. In doing so, the resulting kwarg value
for `model_metadata` was a dict and not an instance of `ModelMetadata`, causing
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/776.
To fix this, we make `get_on_behalf` handle `None` values inside it, so callers
don't have to worry about such minutiae.